Ignore GPOS kerning if the returned value is 0
Some fonts feature both a GPOS & kern table,
but stb may not always be able to parse every GPOS format out there,
which leads stbtt__GetGlyphGPOSInfoAdvance to return 0.

Assume that if the GPOS table exists, it's not just for the sake of
putting a 0 in there, which allows us to assume that 0 more or less
always means we failed to handle the actual GPOS data.
In which case, prefer the kern table (if it exists), instead of using
that 0 and ignoring the kern table.

c.f., https://github.com/nothings/stb/pull/782#issuecomment-512869308